+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ What's new in version 0.6
   Parameters ``other_dots_color`` and ``shading_color`` were added.
   ``facecolor`` will now default to white if
   ``matplotlib.rcParams['axes.facecolor']`` is dark. (:issue:`138`)
+- Added `from_indicators` to allow yet another data input format. This allows
+  category membership to be easily derived from a DataFrame, such as when
+  plotting missing values in the columns of a DataFrame. (:issue:`143`)
 
 What's new in version 0.5
 -------------------------